requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
scikit-learn
joblib
google-generativeai
python-dotenv
argon2_cffi
